+++ ICD.Connect.Audio.Biamp/Controls/Dialing/Telephone/TiDialingDeviceControl.cs
@@ -423,38 +423,53 @@ namespace ICD.Connect.Audio.Biamp.Controls.Dialing.Telephone
 		}
 
 		/// <summary>
-		/// Setup the source callbacks.
+		/// Subscribe to the source callbacks.
 		/// </summary>
 		/// <param name="source"></param>
 		private void Subscribe(ThinConferenceSource source)
 		{
-			source.OnAnswerCallback += AnswerCallback;
-			source.OnHoldCallback += HoldCallback;
-			source.OnResumeCallback += ResumeCallback;
-			source.OnHangupCallback += HangupCallback;
-			source.OnSendDtmfCallback += SendDtmfCallback;
+			source.AnswerCallback += AnswerCallback;
+			source.HoldCallback += HoldCallback;
+			source.ResumeCallback += ResumeCallback;
+			source.SendDtmfCallback += SendDtmfCallback;
+			source.HangupCallback += HangupCallback;
 		}
 
 		/// <summary>
-		/// Remove the source callbacks.
+		/// Unsubscribe from the source callbacks.
 		/// </summary>
 		/// <param name="source"></param>
 		private void Unsubscribe(ThinConferenceSource source)
 		{
-			source.OnAnswerCallback -= AnswerCallback;
-			source.OnHoldCallback -= HoldCallback;
-			source.OnResumeCallback -= ResumeCallback;
-			source.OnHangupCallback -= HangupCallback;
-			source.OnSendDtmfCallback -= SendDtmfCallback;
+			source.AnswerCallback = null;
+			source.HoldCallback = null;
+			source.ResumeCallback = null;
+			source.SendDtmfCallback = null;
+			source.HangupCallback = null;
 		}
 
-		private void SendDtmfCallback(object sender, StringEventArgs stringEventArgs)
+		private void AnswerCallback(ThinConferenceSource sender)
 		{
-			foreach (char digit in stringEventArgs.Data)
+			m_TiControl.Answer();
+		}
+
+		private void HoldCallback(ThinConferenceSource sender)
+		{
+			SetHold(true);
+		}
+
+		private void ResumeCallback(ThinConferenceSource sender)
+		{
+			SetHold(false);
+		}
+
+		private void SendDtmfCallback(ThinConferenceSource sender, string data)
+		{
+			foreach (char digit in data)
 				m_TiControl.Dtmf(digit);
 		}
 
-		private void HangupCallback(object sender, EventArgs eventArgs)
+		private void HangupCallback(ThinConferenceSource sender)
 		{
 			// Ends the active call.
 			m_TiControl.End();
@@ -464,21 +479,6 @@ namespace ICD.Connect.Audio.Biamp.Controls.Dialing.Telephone
 			m_TiControl.SetHookState(TiControlStatusBlock.eHookState.OffHook);
 			m_TiControl.SetHookState(TiControlStatusBlock.eHookState.OnHook);
 			SetHold(false);
-		}
-
-		private void ResumeCallback(object sender, EventArgs eventArgs)
-		{
-			SetHold(false);
-		}
-
-		private void HoldCallback(object sender, EventArgs eventArgs)
-		{
-			SetHold(true);
-		}
-
-		private void AnswerCallback(object sender, EventArgs eventArgs)
-		{
-			m_TiControl.Answer();
 		}
 
 		#endregion