+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
  - Moved volume controls, volume points, volume repeaters and more from other projects
+### Changed
+ - Updated to Callbacks for thinconferencesource 
 
 ## [3.1.0] - 2018-05-02
 ### Added